refactor(codegen): Get rid of legacy idl type
