+++ internal/code-gen/script-wrappers/model/es_operation.go
@@ -1,6 +1,9 @@
 package model
 
 import (
+	"strings"
+
+	htmlelements "github.com/gost-dom/code-gen/html-elements"
 	"github.com/gost-dom/code-gen/script-wrappers/configuration"
 	"github.com/gost-dom/webref/idl"
 )
@@ -8,7 +11,6 @@ import (
 type ESOperation struct {
 	Name                 string
 	NotImplemented       bool
-	LegacyRetType        idl.RetType
 	RetType              idl.Type
 	HasError             bool
 	CustomImplementation bool
@@ -28,13 +30,29 @@ func (op ESOperation) HasResult() bool {
 	return op.RetType.Name != "undefined"
 }
 
+func IsNodeType(typeName string) bool {
+	loweredName := strings.ToLower(typeName)
+	switch loweredName {
+	case "node":
+		return true
+	case "document":
+		return true
+	case "documentfragment":
+		return true
+	}
+	if strings.HasSuffix(loweredName, "element") {
+		return true
+	}
+	return false
+}
+
 func encoderForIDLType(t idl.Type) string {
 	converter := "to"
 	if t.Kind == idl.KindSequence {
 		converter += "Sequence"
 		t = *t.TypeParam
 	}
-	if t.Nullable {
+	if t.Nullable && !htmlelements.IdlType(t).Nillable() {
 		converter += "Nullable"
 	}
 	converter += IdlNameToGoName(t.Name)
@@ -49,5 +67,5 @@ func (o ESOperation) Encoder() string {
 }
 
 func (o ESOperation) RetTypeName() string {
-	return IdlNameToGoName(o.LegacyRetType.TypeName)
+	return o.RetType.Name
 }
